+++ naga/interpolation/radial_point_method.cuh
@@ -74,8 +74,8 @@ class radial_point_method {
   public:
     radial_point_method() = default;
 
-    static size_t get_scratchpad_size(
-        size_t query_size,
+    static uint get_scratchpad_size(
+        uint query_size,
         uint support_size,
         uint dimensions,
         uint group_size = 1
@@ -86,7 +86,7 @@ class radial_point_method {
                 "naga::interpolation::radial_point_method::"
             );
         }
-        size_t mem_per_group
+        uint mem_per_group
             = detail::radial_point_method::get_scratchpad_size_per_group<T>(
                 support_size,
                 dimensions,
@@ -100,7 +100,7 @@ class radial_point_method {
         class ShapeFunctionType = default_shape_function<PointMapType>>
     static sclx::array<T, 2> compute_weights(
         sclx::array<const T, 2> source_points,
-        sclx::array<const size_t, 2> interpolating_indices,
+        sclx::array<const uint, 2> interpolating_indices,
         const PointMapType& query_points,
         const T& approx_particle_spacing,
         uint group_size                         = 1,
@@ -123,7 +123,7 @@ class radial_point_method {
         class ShapeFunctionType = default_shape_function<PointMapType>>
     static radial_point_method<T> create_interpolator(
         sclx::array<const T, 2> source_points,
-        sclx::array<size_t, 2> interpolating_indices,
+        sclx::array<uint, 2> interpolating_indices,
         const PointMapType& query_points,
         const T& approx_particle_spacing,
         uint group_size                         = 1,
@@ -304,7 +304,7 @@ class radial_point_method {
     void load(Archive& ar) {
         ar(group_size_, source_points_size_);
         sclx::deserialize_array(ar, weights_);
-        sclx::array<size_t, 2> indices;
+        sclx::array<uint, 2> indices;
         sclx::deserialize_array(ar, indices);
         indices_ = indices;
     }
@@ -316,9 +316,9 @@ class radial_point_method {
     // usually use the nearest neighbors of their associated, integrated point
     // as their source points.
     uint group_size_{};
-    size_t source_points_size_{};
+    uint source_points_size_{};
     sclx::array<T, 2> weights_{};
-    sclx::array<size_t, 2> indices_{};
+    sclx::array<uint, 2> indices_{};
 
     using matrix_type
         = naga::linalg::matrix<T, naga::linalg::storage_type::sparse_csr>;
